Change default api client request body serialization
This commit makes a change in the default api client - request body
serialization. The earlier code has a conversion for the request
body to convert to json before triggering the APIs, without checking
the content type of the request. This would make incorrect API calls
if the request body is not of type JSON. This commit checks if the
request body content is of type json and only convert it. If not,
then send in the body unconverted, leaving the initial input as is.