Delete persistence attributes without checking setter
This commit fixes the delete persistence attributes method call in attributes manager
to call the delete item method even if the persistence attributes set boolean is not
available. Since delete item is idempotent and there might be cases where skill devs
might have to delete the attributes even before retrieving them, this needs to be
fixed.

The commit also adds the documentation for delete attributes call.